modify _generate_extreme_vectors value to int

+++ application/services/pair_generation/extreme_vectors_strategy.py
@@ -44,7 +44,7 @@ class ExtremeVectorsStrategy(PairGenerationStrategy):
         """
         extremes = []
         for i in range(vector_size):
-            extreme = np.zeros(vector_size)
+            extreme = np.zeros(vector_size, dtype=int)
             extreme[i] = 100
             extremes.append(extreme)
         return extremes
